Res-centric dunbrack (#270)
Implements the dunbrack energy term in a rec-centric way.

This uses some of the existing parameter loading code, and copies of the potential functions. Unlike the existing implementation, this new implementation launches a single kernel for each residue and does all calculations inside of that. 

This term uses a unique system for batching the module parameters inside a wrapper on the python side.