Rework elec count-pair representative lookup for residues w/o yaml entries
No need to raise an exception because a residue type has not specified
a count-pair representative. Water, e.g., does not need to: none of its
atoms are far enough apart for charge-splitting in intra-molecular
interactions to be an issue. No need to put a "no-entry" entry into
the count-pair-representatives list for water just to avoid an exception.