Raise an exception if a NaN coordinate given for an atom marked as present
If the user wants to mark which atoms are present and which are absent,
which is a fine thing to do, and the user has marked one or more
atoms as present but given a coordinate for that atom of NaN, then
tmol cannot proceed. It is perfectly accecptible for the user to
omit the atom_is_present argument in the input to
pose_stack_from_canonical_form, in which case, the user should
use NaN to designate when coordinates are not provided for an atom.