Construct id, frame_x, frame_y, and frame_z tensors w/ C++
The tricky part here is that we had to first construct the
child_list and child_list_span tensors that describe the
descendants of each atom so that we could then traverse
up and down the parent tree and thus correct the
frame_x, _y, and _z tensors.

TO DO: replace bubble sort with mgpu's segmented sort
to better handle cases where we might have a large number
of descendants of a single node (e.g. the root!) but to
still produce a deterministic kintree in spite of
the non-determinism that the atomic-increment based ordering
that I have currently.