Constraints (#305)
Adding the constraint energy term.

The constraint energy term allows you to define custom score functions in Python, alongside sets of atoms that the this score function will be evaluated on. Several functions are built-in to tmol.

The atom sets for each constraint are specified in a ConstraintSet object which is attached to the PoseStack.