Add virtual residue and a test to ensure they do not change the score
A litle tricky part here is that if a virtual residue is labeled as
part of a chain, e.g. as the last residue of a chain, then the previous
residue would not have been detected as the C-term residue of the chain.

I have expanded the logic for chain-ending detection to handle that
case, but have not yet added a test to make sure that logic is correct.