Aleaverfay/kinforest from stencil (#299)
Create KinForests on the GPU (mostly).

The user specifies the kinematic system they want in terms of edges in a FoldForest:
- polymer edges for connections through the up & down connection points in polymeric residues (blocks)
- jump edges to connect arbitrary pairs of residues (using their default jump connection atoms), and
- root jump edges that connect a residue by a jump to the explicit virtual root at the origin

For each block type (residue type) with n_conn connections, the kinforest-from-stencil algorithm creates (n_conn + 2)^2 annotations as an (n_conn + 2) x (n_conn + 2) matrixing of input-connection sources x first-output-connection sinks. In addition to the n_conn connection point entries/first-exits, there are two additional input sources: jumps and root jumps; and there are two additional first output sinks: jumps and "no exit" (when the block is a leaf in the fold forest). For each of these combinations of input source and output sink, the algorithm annotates the block type with a set of _scan-path segments_ (SPSs) representing how the generational-segmented-scan algorithm will compute the coordinates for the atoms in a block of this block type in the context of the full structure; the assembly of the KinForest will join SPSs together to create _scan paths_ (SPs) that will calculate coordinates in the same segmented-scan generation.

The KinForest construction algorithm begins on the CPU with a depth-first traversal of the edges in the fold forest, then a leaf-to-root traversal to select the "first descendant" edges for each branching node (by minimizing the number of generations in the generational segmented scan algorithm), a root-to-leaf traversal to assign "edge delays" (how many generations will be traversed before any SPSs for the blocks on that edge are folded), and a topological sort to construct a total-ordering of the edges based on a partial ordering of 3-tuples of a) edge delay, b) breadth-first traversal order and c) the position of the edge in the unbreaking path of first descendants that will build it. Thus if the polymer edge (0, 10) has as its first descendant the jump (10, 100) and this edge has as its first descendant the polymer edge (100,200), then these three edges would end up "adjacent" in the topological sort. Thus we will put the atoms for residue 100 next to the atoms for residue 10 when constructing the scan path that begins at residue 0.

At this point, the computation moves to the GPU. From here, the GPU shuffles the SPS stencils for each block into the right position in the "nodes" tensor. To compute where each SPS is to be written, the algorithm must first compute offsets into the nodes tensor for each SPS by running scan on the number of atoms each SPS contains. The position where each SPS must write down its number of atoms in the number-of-atoms-for-SPS tensor is given by an offset based on the number of SPSs each residue has (which will be computed using scan), and the position where each residue should write down the number of SPSs it has (per generation) is given by an offset based on the number of residues that each edge builds (again computed using scan) where the edges write down their counts in the topological-sort order. Thus there are three scan steps to compute offsets. Offsets for offsets for offsets.

This PR removes the remnants of the old PackedResidueSystem which was preserved until now only to keep the unit tests running on the kinematics code; now that this algorithm is fully operational, the PackedResidueSystem can be removed.