Unpin torch from 2.6 (#320)
Torch 2.5.0, 2.5.1, and 2.7 also seem to work just fine for tmol; Torch 2.4 does not work.

Also updating the code relying on cattrs.structure

The newest versions of cattrs (>=24.1.0) throw errors that previously were
not errors or that could be worked around or something so that now
the code in tmol/chemical/restypes.py produces errors that weren't
there before.

One small issue: numpy.array cannot be used as a type name for attrs
and this produces a weird complaint out of cattrs. numpy.array is
for creating ndrarrays; ndarrays are the types, array is the function